Merged PR 12540491: Rust Bins: Bake the full PDB path into the Rust binaries
Rust Bins: Bake the full PDB path into the Rust binaries

Starting with rustc 1.78 and later, the compiler (rustc) will automatically pass
the flag /PDBALTPATH:%_PDB% to the linker (rust-lld). Note that %_PDB% is not
read from the environment; instead, The linker will replace %_PDB% with the base
name of the module being linked. This is not the desired behavior for us as it
will break WinDBG in multiple ways, as we need the full path to the PDB file.
Fortunately, if the flag exists but is empty, LLD will populate it with the full
path of the PDB file!

Ref:
https://github.com/rust-lang/rust/pull/121297
https://github.com/rust-lang/rust/blob/75530e9f72a1990ed2305e16fd51d02f47048f12/compiler/rustc_codegen_ssa/src/back/linker.rs#L1049-L1056
https://github.com/llvm/llvm-project/blob/9ed772cecc23f5f5b060720399b010275bbb7457/lld/COFF/Driver.cpp#L2433-L2446

Signed-off-by: Vineel Kovvuri[MSFT] <vineelko@microsoft.com>

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request ensures that the full PDB path is baked into Rust binaries to avoid issues with WinDBG.
- Updated `/.cargo/config.toml` to include the `-C link-arg=/PDBALTPATH:` flag for various targets, ensuring the full path to the PDB file is used.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->