.gitattributes: Add file (#25)
## Description

Unsets the text attribute on all files to prevent their line endings
from being normalized.

This was previously accomplished solely by instructing users to set
core.autocrlf to false. However, that does not translate to pipelines
and setting in a pipeline requires the system level git configuration to
be modified prior to checkout. Git configs cannot be checked into to a
repo and automatically used for security reasons.

This is a simple way to check the change into the repo so it applies for
all users including pipelines to prevent line ending conversion.

This makes the repo consistent with other Patina repos.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Copied from other repos.

## Integration Instructions

N/A

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>