Add enable_debugger feature to Q35/SBSA
Cargo bloat has shown that instantiating the debugger, even
when not in use, adds between 60k and 200k to the file size,
depending on platform. This creates a new model to save that
space by adding a feature (which can be set via cmd line) to
enable the debugger, saving the file space if not used.

Before this, the debugger needed to be enabled via a code change
anyway, so a build was already required to enable it.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>