Makefile.toml: Unify Make Arg Handling
Commit e3dd8615044ad2e9dafe9801840fece4cf40f2e9 was tested
using patina-qemu's build and run script, which passes arguments
in a different fashion than directly running cargo make. As a
result, using that script worked with the Makefile.toml changes
introduced in the above commit, but running the same commands
locally could fail to produce the desired result. This commit
ensures that the cargo make syntax is respected and a separate PR
is fixing patina-qemu's build and run script so it passed arguments
in the same order as cargo make naturally does.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>