Update to Patina v12.1.0 (#54)
## Description

Integrates the latest patina release. The main integration change needed
is to update the path for sample components used in the Q35 platform
build.

The version of `qemu_dxe_core` is updated to 2.0.1. Only the patch
version is updated since this release is not expected to require any
specific integration work needed by a platform using the binary.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

- `cargo make all`
- QEMU Q35 boot to EFI shell
- QEMU SBSA boot is in progress

## Integration Instructions

- Platform firmware is not expected to need any changes to integrate
this update.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>